Chore: Fixed a spurious CI failure (#5685)
### Description
I recently stumbled on this random Mac failure:

https://github.com/dafny-lang/dafny/actions/runs/10359400713/job/28675744622?pr=5684
After careful investigation of the stack trace, it looks like the static
constructor of `DafnyFile` did not run before
`OptionRegistry.CheckOptionsAreKnown(AllOptions);`, resulting in the
option `DoNotVerifyDependencies` to not be registered. This PR fixes
that by forcing the execution of the static `DafnyFile` constructor,
guaranteed to not run if that constructor was already called.
https://learn.microsoft.com/en-us/dotnet/api/system.runtime.compilerservices.runtimehelpers.runclassconstructor?view=net-8.0

### How has this been tested?
I can't think of a way this change can be tested, the failure happens
extremely rarely. Please make yourself convinced that this PR solves the
issue given in the link.

<small>By submitting this pull request, I confirm that my contribution
is made under the terms of the [MIT
license](https://github.com/dafny-lang/dafny/blob/master/LICENSE.txt).</small>